fix: files_scanned u32→u64 prevents silent truncation (closes #137) (#138)
* fix: apply Option B hotfix and Option A u64 migration for files_scanned

* chore: apply rustfmt to overflow_protection test

* fix: update diffguard-core test fixtures for files_scanned u64 migration

* fix: remove $id from JSON schemas (jsonschema crate compatibility)

* fix: remove placeholder assertion from overflow test (clippy)

* fix: resolve clippy bool assertion warnings in diffguard-types tests

* chore: re-trigger CI with updated PR body

+++ crates/diffguard-core/tests/properties.rs
@@ -133,7 +133,7 @@ fn arb_diff_meta() -> impl Strategy<Value = DiffMeta> {
         arb_identifier(), // head
         0u32..10,         // context_lines
         arb_scope(),      // scope
-        0u32..100,        // files_scanned
+        0u64..100,        // files_scanned
         0u32..1000,       // lines_scanned
     )
         .prop_map(
@@ -642,7 +642,7 @@ mod unit_tests {
                 head: "HEAD".to_string(),
                 context_lines: u32::MAX,
                 scope: Scope::Added,
-                files_scanned: u32::MAX,
+                files_scanned: u64::MAX,
                 lines_scanned: u32::MAX,
             },
             findings: vec![],
